Auto reload webcam images;

+++ panoptes/environment/webcams.py
@@ -1,4 +1,4 @@
-import os
+import os, os.path
 import sys
 import subprocess
 import time
@@ -136,8 +136,8 @@ class Webcams(object):
                 self.logger.debug("Image captured for {}".format(webcam.get('name')))
 
                 # Symlink the latest image
-                os.unlink(static_out_file)
-                os.unlink(static_thumbnail_file)
+                if os.path.exists(static_out_file): os.unlink(static_out_file)
+                if os.path.exists(static_thumbnail_file): os.unlink(static_thumbnail_file)
 
                 os.symlink(out_file, static_out_file)
                 os.symlink(thumbnail_file, static_thumbnail_file)
